refactor backend into backendshell and backendchannel to allos multiple backend instances
