+++ java/shell.scala
@@ -1,0 +1,50 @@
+package sparklyr
+
+import java.io.{DataInputStream, DataOutputStream}
+import java.io.{File, FileOutputStream, IOException}
+import java.net.{InetAddress, InetSocketAddress, ServerSocket, Socket}
+import java.util.concurrent.TimeUnit
+
+import io.netty.bootstrap.ServerBootstrap
+import io.netty.channel.{ChannelFuture, ChannelInitializer, EventLoopGroup}
+import io.netty.channel.nio.NioEventLoopGroup
+import io.netty.channel.socket.SocketChannel
+import io.netty.channel.socket.nio.NioServerSocketChannel
+import io.netty.handler.codec.LengthFieldBasedFrameDecoder
+import io.netty.handler.codec.bytes.{ByteArrayDecoder, ByteArrayEncoder}
+
+import org.apache.spark.SparkConf
+import org.apache.spark.SparkContext
+import org.apache.spark.sql.hive.HiveContext
+
+import scala.util.Try
+
+object Shell {
+  private[this] var backend: Backend = null
+
+  def main(args: Array[String]): Unit = {
+    if (args.length > 4 || args.length < 2) {
+      System.err.println(
+        "Usage: Backend port id [--service] [--remote]\n" +
+        "  port:      port the gateway will listen to\n" +
+        "  id:        arbitrary numeric identifier for this backend session\n" +
+        "  --service: prevents closing the connection from closing the backen\n" +
+        "  --remote:  allows the gateway to accept remote connections\n"
+      )
+
+      System.exit(-1)
+    }
+
+    val port = args(0).toInt
+    val sessionId = args(1).toInt
+    val isService = args.contains("--service")
+    val isRemote = args.contains("--remote")
+
+    backend = new Backend()
+    backend.init(port, sessionId, isService, isRemote, false)
+  }
+
+  def getBackend(): Backend = {
+    backend
+  }
+}
